refactor: reorder hook (#6483)
