+++ frontend/src/views/Collection/components/CollectionActions/components/MoreDropdown/index.tsx
@@ -3,7 +3,7 @@ import { Collection, Dataset } from "src/common/entities";
 import Menu from "./components/Menu";
 import { DeleteCollectionFn } from "src/views/Collection/components/CollectionActions";
 import { ButtonIcon } from "src/views/Collection/components/CollectionActions/components/MoreDropdown/style";
-import { ReorderAction } from "src/views/Collection/hooks/useReorderMode";
+import { ReorderAction } from "src/views/Collection/hooks/useReorderMode/useReorderMode";
 
 interface Props {
   collection: Collection;
