+++ frontend/src/views/Collection/index.tsx
@@ -34,7 +34,7 @@ import CollectionActions from "src/views/Collection/components/CollectionActions
 import {
   REORDER_MODE,
   useReorderMode,
-} from "src/views/Collection/hooks/useReorderMode";
+} from "src/views/Collection/hooks/useReorderMode/useReorderMode";
 
 const Collection: FC = () => {
   const isCurator = get(FEATURES.CURATOR) === BOOLEAN.TRUE;
